server/Auth/requirements.txt
asgiref==3.8.1
boto3==1.35.10
botocore==1.35.10
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
idna==3.8
jmespath==1.0.1
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.2
shortuuid==1.0.13
six==1.16.0
sqlparse==0.5.1
urllib3==2.2.2
